+++ Sources/Operation+Injection.swift
@@ -49,14 +49,29 @@ public struct InjectedInputRequirements: OptionSet {
 }
 
 extension OutputHaving where Self: AdvancedOperation {
+
   /// Creates a new operation that passes the output of `self` into the given `AdvancedOperation`
   ///
   /// - Parameters:
   ///   - operation: The operation that needs the output of `self` to generate an output.
   ///   - requirements: A list of options that the injected input must satisfy.
   /// - Returns: Returns an *adapter* operation which passes the output of `self` into the given `AdvancedOperation`.
-  public func inject<E: InputHaving & AdvancedOperation>(into operation: E, requirements: InjectedInputRequirements = [.notOptional, .successful]) -> AdvancedBlockOperation where Output == E.Input {
+  public func inject<E: InputHaving & AdvancedOperation>(into operation: E,
+                                                         requirements: InjectedInputRequirements = [.notOptional, .successful]) -> AdvancedBlockOperation where Output == E.Input {
     return AdvancedOperation.injectOperation(self, into: operation, requirements: requirements)
+  }
+
+  /// Creates a new operation that passes, once transformed, the output of `self` into the given `AdvancedOperation`.
+  ///
+  /// - Parameters:
+  ///   - operation: The operation that needs the transformed output of `self` to generate an output.
+  ///   - requirements: A list of options that the injected input must satisfy.
+  ///   - transform: Closure to transform the output of `self` into a valid `input` for the next operation.
+  /// - Returns: Returns an *adapter* operation which passes the transformed output of `self` into the given `AdvancedOperation`.
+  public func inject<E: InputHaving & AdvancedOperation>(into operation: E,
+                                                         requirements: InjectedInputRequirements = [.notOptional, .successful],
+                                                         transform: @escaping (Output?) -> E.Input?) -> AdvancedBlockOperation {
+    return AdvancedOperation.injectOperation(self, into: operation, requirements: requirements, transform: { transform($0) })
   }
 }
 
@@ -77,21 +92,7 @@ extension AdvancedOperation {
   class func injectOperation<F: OutputHaving & AdvancedOperation, G: InputHaving & AdvancedOperation>(_ outputOperation: F, into inputOpertion: G, requirements: InjectedInputRequirements = []) -> AdvancedBlockOperation where F.Output == G.Input {
     let adapterOperation = AdvancedBlockOperation { [unowned outputOperation = outputOperation, unowned inputOpertion = inputOpertion] complete in
 
-      let error: NSError? = {
-        if requirements.contains(.notOptional) && outputOperation.output == nil {
-          return AdvancedOperationError.executionCancelled(message: "The injectable input is nil and it doesn't satisfy the requirements (\(requirements).")
-        }
-
-        if requirements.contains(.successful), !outputOperation.errors.isEmpty {
-          return AdvancedOperationError.executionCancelled(message: "The injectable operation contains errors and it doesn't satisfy the requirements (\(requirements).")
-        }
-
-        if requirements.contains(.noCancellation), outputOperation.isCancelled {
-          return AdvancedOperationError.executionCancelled(message: "The injectable operation is cancelledand it doesn't satisfy the requirements (\(requirements).")
-        }
-
-        return nil
-      }()
+    let error = AdvancedOperation.evaluateOutputOperation(outputOperation, forRequirements: requirements)
 
       inputOpertion.input = outputOperation.output
       if let error = error {
@@ -104,5 +105,50 @@ extension AdvancedOperation {
     inputOpertion.addDependency(adapterOperation)
 
     return adapterOperation
+  }
+
+  /// Creates an *adapter* operation which passes, once transformed, the output from the `outputOperation` into the input of the `inputOpertion`.
+  ///
+  /// - Parameters:
+  ///   - outputOperation: The operation whose output is needed by the `inputOperation`.
+  ///   - inputOpertion: The operation who needs needs, as input, the output of the `inputOperation`.
+  ///   - requirements: A set of `InjectedInputRequirements`.
+  ///   - transform: Closure to transform the output of `self` into a valid `input` for the next operation.
+  /// - Returns: Returns an *adapter* operation which passes the transformed output from the `outputOperation` into the input of the `inputOpertion`,
+  /// and builds dependencies so the outputOperation runs first, then the adapter, then inputOpertion.
+  /// - Note: The client is still responsible for adding all three blocks to a queue.
+  // swiftlint:disable:next line_length
+  class func injectOperation<F: OutputHaving & AdvancedOperation, G: InputHaving & AdvancedOperation>(_ outputOperation: F, into inputOpertion: G, requirements: InjectedInputRequirements = [], transform: @escaping (F.Output?) -> G.Input?) -> AdvancedBlockOperation {
+    let adapterOperation = AdvancedBlockOperation { [unowned outputOperation = outputOperation, unowned inputOpertion = inputOpertion] complete in
+
+      let error = AdvancedOperation.evaluateOutputOperation(outputOperation, forRequirements: requirements)
+
+      inputOpertion.input = transform(outputOperation.output)
+      if let error = error {
+        inputOpertion.cancel(error: error)
+      }
+      complete([])
+    }
+
+    adapterOperation.addDependency(outputOperation)
+    inputOpertion.addDependency(adapterOperation)
+
+    return adapterOperation
+  }
+
+  private class func evaluateOutputOperation<F: OutputHaving & AdvancedOperation>(_ outputOperation: F, forRequirements requirements: InjectedInputRequirements) -> NSError? {
+      if requirements.contains(.notOptional) && outputOperation.output == nil {
+        return AdvancedOperationError.executionCancelled(message: "The injectable input is nil and it doesn't satisfy the requirements (\(requirements).")
+      }
+
+      if requirements.contains(.successful), !outputOperation.errors.isEmpty {
+        return AdvancedOperationError.executionCancelled(message: "The injectable operation contains errors and it doesn't satisfy the requirements (\(requirements).")
+      }
+
+      if requirements.contains(.noCancellation), outputOperation.isCancelled {
+        return AdvancedOperationError.executionCancelled(message: "The injectable operation is cancelledand it doesn't satisfy the requirements (\(requirements).")
+      }
+
+      return nil
   }
 }